+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/pnl/ProfitLossRatioCriterion.java
@@ -30,23 +30,42 @@ import org.ta4j.core.analysis.criteria.AbstractAnalysisCriterion;
 import org.ta4j.core.num.Num;
 
 /**
- * Gross return criterion (with commissions).
- *
- * <p>
- * The gross return of the provided {@link Position position(s)} over the
- * provided {@link BarSeries series}.
+ * Ratio Profit Loss Criterion = Average gross profit (with commissions) /
+ * Average gross loss (with commissions)
  */
-public class GrossReturnCriterion extends AbstractAnalysisCriterion {
+public class ProfitLossRatioCriterion extends AbstractAnalysisCriterion {
+
+    private AverageProfitCriterion averageProfitCriterion = new AverageProfitCriterion();
+    private AverageLossCriterion averageLossCriterion = new AverageLossCriterion();
 
     @Override
     public Num calculate(BarSeries series, Position position) {
-        return calculateProfit(series, position);
+        Num averageProfit = averageProfitCriterion.calculate(series, position);
+        if (averageProfit.isZero()) {
+            // only loosing positions means a ratio of 0
+            return series.numOf(0);
+        }
+        Num averageLoss = averageLossCriterion.calculate(series, position);
+        if (averageLoss.isZero()) {
+            // only winning positions means a ratio of 1
+            return series.numOf(1);
+        }
+        return averageProfit.dividedBy(averageLoss).abs();
     }
 
     @Override
     public Num calculate(BarSeries series, TradingRecord tradingRecord) {
-        return tradingRecord.getPositions().stream().map(position -> calculateProfit(series, position))
-                .reduce(series.numOf(1), Num::multipliedBy);
+        Num averageProfit = averageProfitCriterion.calculate(series, tradingRecord);
+        if (averageProfit.isZero()) {
+            // only loosing positions means a ratio of 0
+            return series.numOf(0);
+        }
+        Num averageLoss = averageLossCriterion.calculate(series, tradingRecord);
+        if (averageLoss.isZero()) {
+            // only winning positions means a ratio of 1
+            return series.numOf(1);
+        }
+        return averageProfit.dividedBy(averageLoss).abs();
     }
 
     @Override
@@ -54,17 +73,4 @@ public class GrossReturnCriterion extends AbstractAnalysisCriterion {
         return criterionValue1.isGreaterThan(criterionValue2);
     }
 
-    /**
-     * Calculates the gross return of a position (Buy and sell).
-     *
-     * @param series   a bar series
-     * @param position a position
-     * @return the gross return of the position
-     */
-    private Num calculateProfit(BarSeries series, Position position) {
-        if (position.isClosed()) {
-            return position.getGrossReturn(series);
-        }
-        return series.numOf(1);
-    }
 }