+++ ta4j-core/src/main/java/org/ta4j/core/indicators/DistanceFromMAIndicator.java
@@ -1,0 +1,71 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.indicators;
+
+import org.ta4j.core.Bar;
+import org.ta4j.core.BarSeries;
+import org.ta4j.core.num.Num;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+/**
+ * Distance From Moving Average (close - MA)/MA
+ *
+ * @see <a href=
+ *      "https://school.stockcharts.com/doku.php?id=technical_indicators:distance_from_ma">
+ *      https://school.stockcharts.com/doku.php?id=technical_indicators:distance_from_ma
+ *      </a>
+ */
+public class DistanceFromMAIndicator extends CachedIndicator<Num> {
+    private static final Set<Class> supportedMovingAverages = new HashSet<>(
+            Arrays.asList(EMAIndicator.class, DoubleEMAIndicator.class, TripleEMAIndicator.class, SMAIndicator.class,
+                    WMAIndicator.class, ZLEMAIndicator.class, HMAIndicator.class, KAMAIndicator.class,
+                    LWMAIndicator.class, AbstractEMAIndicator.class, MMAIndicator.class));
+    private final CachedIndicator movingAverage;
+
+    /**
+     *
+     * @param series        the bar series {@link BarSeries}.
+     * @param movingAverage the moving average.
+     */
+    public DistanceFromMAIndicator(BarSeries series, CachedIndicator movingAverage) {
+        super(series);
+        if (!(supportedMovingAverages.contains(movingAverage.getClass()))) {
+            throw new IllegalArgumentException(
+                    "Passed indicator must be a moving average based indicator. " + movingAverage.toString());
+        }
+        this.movingAverage = movingAverage;
+    }
+
+    @Override
+    protected Num calculate(int index) {
+        Bar currentBar = getBarSeries().getBar(index);
+        Num closePrice = currentBar.getClosePrice();
+        Num maValue = (Num) movingAverage.getValue(index);
+        return (closePrice.minus(maValue)).dividedBy(maValue);
+    }
+}